[hw-94] - Add role-based authorization to artists, albums, and tracks; - Implement only admin deletion for artists, albums, and tracks.

+++ src/albums/albums.controller.ts
@@ -2,11 +2,14 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
   Query,
+  Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
@@ -15,6 +18,9 @@ import { Album, AlbumDocument } from '../schemas/album.schema';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { CreateAlbumDto } from './create.album.dto';
 import { albumImage } from '../middleware/multer';
+import { TokenAuthGuard } from '../token-auth/token-auth.guard';
+import { UserDocument } from '../schemas/user.schema';
+import { Request } from 'express';
 
 @Controller('albums')
 export class AlbumsController {
@@ -37,6 +43,7 @@ export class AlbumsController {
   }
 
   @Post()
+  @UseGuards(TokenAuthGuard)
   @UseInterceptors(FileInterceptor('image', { storage: albumImage }))
   create(
     @UploadedFile() file: Express.Multer.File,
@@ -52,7 +59,17 @@ export class AlbumsController {
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  @UseGuards(TokenAuthGuard)
+  async remove(
+    @Param('id') id: string,
+    @Req() req: Request<{ user: UserDocument }>,
+  ) {
+    const user = req.user as UserDocument;
+
+    if (user.role !== 'admin') {
+      throw new ForbiddenException('Only administrators can delete albums.');
+    }
+
     await this.albumModel.findByIdAndDelete(id);
     return { message: 'Album was deleted successfully' };
   }
