+++ src/artists/artists.controller.ts
@@ -2,10 +2,13 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
+  Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
@@ -14,6 +17,9 @@ import { Artist, ArtistDocument } from '../schemas/artist.schema';
 import { CreateArtistDto } from './create.artist.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { artistImage } from '../middleware/multer';
+import { TokenAuthGuard } from '../token-auth/token-auth.guard';
+import { UserDocument } from '../schemas/user.schema';
+import { Request } from 'express';
 
 @Controller('artists')
 export class ArtistsController {
@@ -32,6 +38,7 @@ export class ArtistsController {
   }
 
   @Post()
+  @UseGuards(TokenAuthGuard)
   @UseInterceptors(FileInterceptor('image', { storage: artistImage }))
   create(
     @UploadedFile() file: Express.Multer.File,
@@ -46,7 +53,17 @@ export class ArtistsController {
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  @UseGuards(TokenAuthGuard)
+  async remove(
+    @Param('id') id: string,
+    @Req() req: Request<{ user: UserDocument }>,
+  ) {
+    const user = req.user as UserDocument;
+
+    if (user.role !== 'admin') {
+      throw new ForbiddenException('Only administrators can delete artists');
+    }
+
     await this.artistModel.findByIdAndDelete(id);
     return { message: 'Artist was deleted successfully' };
   }
