+++ src/tracks/tracks.controller.ts
@@ -2,15 +2,21 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
   Query,
+  Req,
+  UseGuards,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateTrackDto } from './create.track.dto';
 import { Track, TrackDocument } from '../schemas/track.schema';
+import { TokenAuthGuard } from '../token-auth/token-auth.guard';
+import { UserDocument } from '../schemas/user.schema';
+import { Request } from 'express';
 
 @Controller('tracks')
 export class TracksController {
@@ -32,6 +38,7 @@ export class TracksController {
     return this.trackModel.find({ _id: id });
   }
 
+  @UseGuards(TokenAuthGuard)
   @Post()
   async create(@Body() trackDto: CreateTrackDto) {
     const track = new this.trackModel({
@@ -45,7 +52,17 @@ export class TracksController {
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  @UseGuards(TokenAuthGuard)
+  async remove(
+    @Param('id') id: string,
+    @Req() req: Request<{ user: UserDocument }>,
+  ) {
+    const user = req.user as UserDocument;
+
+    if (user.role !== 'admin') {
+      throw new ForbiddenException('Only administrators can delete tracks');
+    }
+
     await this.trackModel.findByIdAndDelete(id);
     return { message: 'Track was deleted successfully' };
   }
